requirements.txt
Adafruit-Blinka==8.47.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-typing==1.11.0
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11
Adafruit_DHT==1.4.0
bidict==0.23.1
binho-host-adapter==0.1.6
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-SocketIO==5.3.6
h11==0.14.0
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
pyftdi==0.55.4
pyserial==3.5
python-engineio==4.9.0
python-socketio==5.11.1
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
simple-websocket==1.0.0
spidev==3.6
sysv-ipc==1.1.0
typing_extensions==4.12.2
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
